+++ command/du.go
@@ -111,7 +111,7 @@ func (sz Size) Run(ctx context.Context) error {
 	total := sizeAndCount{}
 
 	var merror error
-	excludePatterns := strutil.CreateExcludesFromWildcard(sz.exclude)
+	excludePatterns := CreateExcludesFromWildcard(sz.exclude)
 
 	for object := range client.List(ctx, srcurl, false) {
 		if object.Type.IsDir() || errorpkg.IsCancelation(object.Err) {
@@ -124,7 +124,7 @@ func (sz Size) Run(ctx context.Context) error {
 			continue
 		}
 
-		if strutil.IsURLExcluded(excludePatterns, object.URL.Path) {
+		if isURLExcluded(excludePatterns, object.URL.Path) {
 			continue
 		}
