+++ command/ls.go
@@ -153,7 +153,7 @@ func (l List) Run(ctx context.Context) error {
 	}
 
 	var merror error
-	excludePatterns := strutil.CreateExcludesFromWildcard(l.exclude)
+	excludePatterns := CreateExcludesFromWildcard(l.exclude)
 
 	for object := range client.List(ctx, srcurl, false) {
 		if errorpkg.IsCancelation(object.Err) {
@@ -166,7 +166,7 @@ func (l List) Run(ctx context.Context) error {
 			continue
 		}
 
-		if strutil.IsURLExcluded(excludePatterns, object.URL.Path) {
+		if isURLExcluded(excludePatterns, object.URL.Path) {
 			continue
 		}
