+++ command/select.go
@@ -15,7 +15,6 @@ import (
 	"github.com/peak/s5cmd/parallel"
 	"github.com/peak/s5cmd/storage"
 	"github.com/peak/s5cmd/storage/url"
-	"github.com/peak/s5cmd/strutil"
 )
 
 var selectHelpTemplate = `Name:
@@ -157,7 +156,7 @@ func (s Select) Run(ctx context.Context) error {
 		}
 	}()
 
-	excludePatterns := strutil.CreateExcludesFromWildcard(s.exclude)
+	excludePatterns := CreateExcludesFromWildcard(s.exclude)
 
 	for object := range objch {
 		if object.Type.IsDir() || errorpkg.IsCancelation(object.Err) {
@@ -175,7 +174,7 @@ func (s Select) Run(ctx context.Context) error {
 			continue
 		}
 
-		if strutil.IsURLExcluded(excludePatterns, object.URL.Path) {
+		if isURLExcluded(excludePatterns, object.URL.Path) {
 			continue
 		}
